backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
distro==1.9.0
dnspython==1.16.0
et-xmlfile==1.1.0
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.41.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.7.2
numpy==1.25.1
openai==1.14.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==3.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.2.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2023.3
Werkzeug==3.0.1
XlsxWriter==3.1.2
